Safari-609.1.11.1/LayoutTests/editing/selection/move-by-word-visually-across-object-element-3-expected.txt

This tests moving left and right by word boundary across an object element. WebKit should not hang.

Initial state:
| <div>
|   "<#selection-caret>שלום"
| <div>
|   <object>
|     border="1"
|     height="50"
|     type="application/x-something"
|     width="50"
|   <br>
|   <div>
|     "עולם"

Move left count: 1:
| <div>
|   "שלום<#selection-caret>"
| <div>
|   <object>
|     border="1"
|     height="50"
|     type="application/x-something"
|     width="50"
|   <br>
|   <div>
|     "עולם"

Move left count: 2:
| <div>
|   "שלום"
| <div>
|   <object>
|     border="1"
|     height="50"
|     type="application/x-something"
|     width="50"
|   <br>
|   <div>
|     "עולם<#selection-caret>"

Move left count: 3:
| <div>
|   "שלום"
| <div>
|   <object>
|     border="1"
|     height="50"
|     type="application/x-something"
|     width="50"
|   <br>
|   <div>
|     "עולם<#selection-caret>"

Move right count: 1:
| <div>
|   "שלום"
| <div>
|   <object>
|     border="1"
|     height="50"
|     type="application/x-something"
|     width="50"
|   <br>
|   <div>
|     "<#selection-caret>עולם"

Move right count: 2:
| <div>
|   "<#selection-caret>שלום"
| <div>
|   <object>
|     border="1"
|     height="50"
|     type="application/x-something"
|     width="50"
|   <br>
|   <div>
|     "עולם"

Move right count: 3:
| <div>
|   "<#selection-caret>שלום"
| <div>
|   <object>
|     border="1"
|     height="50"
|     type="application/x-something"
|     width="50"
|   <br>
|   <div>
|     "עולם"
